--- a/core/Controller/CssController.php
+++ b/core/Controller/CssController.php
@@ -34,6 +34,7 @@ namespace OC\Core\Controller;
 use OC\Files\AppData\Factory;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http;
+use OCP\AppFramework\Http\Attribute\FrontpageRoute;
 use OCP\AppFramework\Http\Attribute\OpenAPI;
 use OCP\AppFramework\Http\FileDisplayResponse;
 use OCP\AppFramework\Http\NotFoundResponse;
@@ -69,6 +70,7 @@ class CssController extends Controller {
 	 * @param string $appName css folder name
 	 * @return FileDisplayResponse|NotFoundResponse
 	 */
+	#[FrontpageRoute(verb: 'GET', url: '/css/{appName}/{fileName}')]
 	public function getCss(string $fileName, string $appName): Response {
 		try {
 			$folder = $this->appData->getFolder($appName);