… apps/user_status/appinfo/routes.php
… apps/user_status/openapi.json
… core/Controller/OCMController.php
… core/Controller/OCSController.php
… core/Controller/PreviewController.php
… core/Controller/ProfileApiController.php
… core/Controller/ProfilePageController.php
… core/Controller/RecommendedAppsController.php
… core/Controller/ReferenceApiController.php
… core/Controller/ReferenceController.php
… core/Controller/SearchController.php
… core/Controller/TextProcessingApiController.php
… core/Controller/TextToImageApiController.php
… core/Controller/TranslationApiController.php
… core/Controller/TwoFactorChallengeController.php
… core/Controller/UnifiedSearchController.php
… core/Controller/UnsupportedBrowserController.php
… core/Controller/UserController.php
… core/Controller/WalledGardenController.php
… core/Controller/WebAuthnController.php
… core/Controller/WellKnownController.php
… core/Controller/WhatsNewController.php
… core/Controller/WipeController.php
… core/openapi.json
… core/routes.php
… lib/composer/composer/autoload_classmap.php
… lib/composer/composer/autoload_static.php
… lib/private/AppFramework/Routing/RouteConfig.php
… lib/private/AppFramework/Routing/RouteParser.php
… lib/private/Route/Router.php