## Summary

Depends on openapi-extractor: https://github.com/nextcloud/openapi-extractor/pull/73
Documentation: https://github.com/nextcloud/documentation/pull/11442

A new routing method that should make it easier for new developers. It avoids having to manually verify the connection between the layers of the routes and the controllers.

The changes to the openapi.json file are annoying but required because the order of the routes changed. You can verify there were no changes using http://play.jd-tool.io (with unordered set option). Since there is no diff we can also be sure I migrated all routes correctly.

I did some benchmarking and would consider the slight drop in performance to be acceptable. It is about 2-5ms which is not noticable and there are many other factors in a remote HTTP connection that cause way more delay (tests were done on localhost). I have some ideas which could be tried to improve the performance a bit (but there is no guarantee that they will).

Only the 99th percentile is really relevant as the longest request can also be a fluke (see in the baseline testing for the capabilities endpoint). All tests were done with 10k requests on the standalone PHP webserver with SQLite and without any caching configured.

The core test is to check for changes in parsing the routes and the test with Talk (which has not been migrated) is to check the impact on apps with many routes. The total number of enabled apps has no performance impact on this new mechanism.

Before:
```log
$ ab -n 10000 -H "OCS-APIRequest: true" http://localhost:8080/ocs/v2.php/cloud/capabilities
This is ApacheBench, Version 2.3 <$Revision: 1903618 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 1000 requests
Completed 2000 requests
Completed 3000 requests
Completed 4000 requests
Completed 5000 requests
Completed 6000 requests
Completed 7000 requests
Completed 8000 requests
Completed 9000 requests
Completed 10000 requests
Finished 10000 requests


Server Software:
Server Hostname:        localhost
Server Port:            8080

Document Path:          /ocs/v2.php/cloud/capabilities
Document Length:        1220 bytes

Concurrency Level:      1
Time taken for tests:   143.307 seconds
Complete requests:      10000
Failed requests:        0
Total transferred:      25310186 bytes
HTML transferred:       12200000 bytes
Requests per second:    69.78 [#/sec] (mean)
Time per request:       14.331 [ms] (mean)
Time per request:       14.331 [ms] (mean, across all concurrent requests)
Transfer rate:          172.48 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:    12   14   4.0     13     354
Waiting:       12   14   4.0     13     353
Total:         12   14   4.0     13     354

Percentage of the requests served within a certain time (ms)
  50%     13
  66%     14
  75%     14
  80%     15
  90%     17
  95%     19
  98%     21
  99%     22
 100%    354 (longest request)

$ ab -n 10000 -H "OCS-APIRequest: true" -H "Authorization: Bearer bgN7T-A4q9p-QTT5C-QcFEK-57wfc" http://localhost:8080/ocs/v2.php/apps/spreed/api/v4/room
This is ApacheBench, Version 2.3 <$Revision: 1903618 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 1000 requests
Completed 2000 requests
Completed 3000 requests
Completed 4000 requests
Completed 5000 requests
Completed 6000 requests
Completed 7000 requests
Completed 8000 requests
Completed 9000 requests
Completed 10000 requests
Finished 10000 requests


Server Software:
Server Hostname:        localhost
Server Port:            8080

Document Path:          /ocs/v2.php/apps/spreed/api/v4/room
Document Length:        7612 bytes

Concurrency Level:      1
Time taken for tests:   260.035 seconds
Complete requests:      10000
Failed requests:        0
Total transferred:      89919180 bytes
HTML transferred:       76120000 bytes
Requests per second:    38.46 [#/sec] (mean)
Time per request:       26.003 [ms] (mean)
Time per request:       26.003 [ms] (mean, across all concurrent requests)
Transfer rate:          337.69 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:    23   26   4.2     24     269
Waiting:       23   26   4.2     24     267
Total:         23   26   4.2     24     269

Percentage of the requests served within a certain time (ms)
  50%     24
  66%     26
  75%     27
  80%     28
  90%     31
  95%     33
  98%     35
  99%     37
 100%    269 (longest request)
```

Now:
```log
$ ab -n 10000 -H "OCS-APIRequest: true" http://localhost:8080/ocs/v2.php/cloud/capabilities
This is ApacheBench, Version 2.3 <$Revision: 1903618 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 1000 requests
Completed 2000 requests
Completed 3000 requests
Completed 4000 requests
Completed 5000 requests
Completed 6000 requests
Completed 7000 requests
Completed 8000 requests
Completed 9000 requests
Completed 10000 requests
Finished 10000 requests


Server Software:
Server Hostname:        localhost
Server Port:            8080

Document Path:          /ocs/v2.php/cloud/capabilities
Document Length:        1220 bytes

Concurrency Level:      1
Time taken for tests:   147.590 seconds
Complete requests:      10000
Failed requests:        0
Total transferred:      25309812 bytes
HTML transferred:       12200000 bytes
Requests per second:    67.76 [#/sec] (mean)
Time per request:       14.759 [ms] (mean)
Time per request:       14.759 [ms] (mean, across all concurrent requests)
Transfer rate:          167.47 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:    13   15   2.5     14      42
Waiting:       12   14   2.5     13      41
Total:         13   15   2.5     14      42

Percentage of the requests served within a certain time (ms)
  50%     14
  66%     14
  75%     15
  80%     16
  90%     18
  95%     20
  98%     22
  99%     24
 100%     42 (longest request)

$ ab -n 10000 -H "OCS-APIRequest: true" -H "Authorization: Bearer bgN7T-A4q9p-QTT5C-QcFEK-57wfc" http://localhost:8080/ocs/v2.php/apps/spreed/api/v4/room
This is ApacheBench, Version 2.3 <$Revision: 1903618 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 1000 requests
Completed 2000 requests
Completed 3000 requests
Completed 4000 requests
Completed 5000 requests
Completed 6000 requests
Completed 7000 requests
Completed 8000 requests
Completed 9000 requests
Completed 10000 requests
Finished 10000 requests


Server Software:
Server Hostname:        localhost
Server Port:            8080

Document Path:          /ocs/v2.php/apps/spreed/api/v4/room
Document Length:        7612 bytes

Concurrency Level:      1
Time taken for tests:   279.000 seconds
Complete requests:      10000
Failed requests:        0
Total transferred:      89920016 bytes
HTML transferred:       76120000 bytes
Requests per second:    35.84 [#/sec] (mean)
Time per request:       27.900 [ms] (mean)
Time per request:       27.900 [ms] (mean, across all concurrent requests)
Transfer rate:          314.74 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.0      0       0
Processing:    24   28   5.3     26     360
Waiting:       24   27   5.2     26     359
Total:         24   28   5.3     27     360

Percentage of the requests served within a certain time (ms)
  50%     27
  66%     27
  75%     28
  80%     29
  90%     34
  95%     38
  98%     39
  99%     42
 100%    360 (longest request)
```

## Checklist

- Code is [properly formatted](https://docs.nextcloud.com/server/latest/developer_manual/digging_deeper/continuous_integration.html#linting)
- [Sign-off message](https://github.com/src-d/guide/blob/master/developer-community/fix-DCO.md) is added to all commits
- [ ] Tests ([unit](https://docs.nextcloud.com/server/latest/developer_manual/app_development/tutorial.html#unit-tests), [integration](https://docs.nextcloud.com/server/latest/developer_manual/app_development/tutorial.html#integration-tests), api and/or acceptance) are included
- [ ] Screenshots before/after for front-end changes
- [x] Documentation ([manuals](https://github.com/nextcloud/documentation/) or wiki) has been updated or is not required
- [ ] [Backports requested](https://github.com/nextcloud/backportbot/#usage) where applicable (ex: critical bugfixes)