@provokateurin 2024-01-15
No idea why OpenAPI changes (seems to use the old version?), but the openapi-extractor PR needs to be merged first anyway.