@nickvergessen 2024-01-15
Since we use the Symfony router under the hood. Would there be an issue extending/using it's attribute as a base? We ship it already: https://github.com/nextcloud/3rdparty/blob/8a78b41c955bb3530a5c25f187e3dc96091dadd7/symfony/routing/Annotation/Route.php

Ref https://symfony.com/doc/current/routing.html

That could even allow us to remove more routing code in the future?