@provokateurin 2024-01-15
> Would there be an issue extending/using it's attribute as a base?

I didn't know about this. Removing routing code sounds good, but this attribute also supports a lot of stuff that we don't support in our routing yet. I don't think that is a good idea :/