@provokateurin 2024-01-15
(I only migrated core and one other app to ensure it works for both since they are handled a bit different in the router)