@ChristophWurst 2024-01-16
Blackfire analysis:
* /login https://blackfire.io/profiles/compare/25ecda33-f01e-46d8-a45f-774c8db2a460/graph
* /ocs/v2.php/cloud/capabilities https://blackfire.io/profiles/compare/9093963f-b6d3-43bb-b695-18a97090dadb/graph

Every comparison is primed with a simple curl to warm caches.
Every test is run with 10 samples to get more accurate numbers.