@provokateurin 2024-01-16
@ChristophWurst thanks for the profiling. When I did my testing I found that even with 1k requests I got very different results every time. At 10k it was quite stable.

To everyone: Do you think the performance hit is acceptable or not?