@provokateurin 2024-02-05
This branch is currently broken if navigate to the root of the instance. Any other URL will work fine. I'm looking into it already.