@provokateurin 2024-02-12
CI failure is fixed, it was just another case of the casing problem with matching legacy routes.